refactored classes to interfaces, put every interface and enum in single file;
