+++ src/app/interfaces/boat-condition.ts
@@ -1,0 +1,7 @@
+export enum BoatCondition {
+  Unknown = 1,
+  Good,
+  Bad,
+  Sinking,
+  'People in water'
+}
