+++ src/app/interfaces/boat-type.ts
@@ -1,0 +1,6 @@
+export enum BoatType {
+  Rubber = 1,
+  Wood,
+  Steel,
+  Other
+}
