+++ src/app/interfaces/location-type.ts
@@ -1,0 +1,4 @@
+export enum LocationType {
+  Case = 1,
+  Vehcile
+}
