+++ src/app/interfaces/status.ts
@@ -1,0 +1,9 @@
+export enum Status {
+  'Need Help' = 1,
+  Critical,
+  Rescued,
+  'Confirmed Target',
+  'Possible Target',
+  Attended,
+  Closed
+}
