+++ src/app/services/locations.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { PouchService } from '../services/pouch.service';
-import { Location } from '../interfaces/case';
+import { Location } from '../interfaces/location';
 
 @Injectable()
 export class LocationsService {
